Program is running, 2 bugs to fix left
The program works fine and ask for a reset at the end of 5 winning
 rounds.
However there are two bugs left that I struggle identifying 
the origin.
    _ Bug 1 is that the first action is always a ghost one meaning 
that after the first choice of the player in the prompt is entered
the prompt pops up instantly after for the player to make a choice
again as if the first never existed. The console shows different
though as it displays both the computer's and the player's choice 
but without counting the points whatever the outcome.
This occurs only for the first prompt, not if you "try again".
    _ Bug 2 is that the tracking of the score is not displayed in the 
console if you open the program in your browser (I use Chrome)
and play directly, it shows if you refresh the page one time before
playing and after it goes on just fine.